containers/app/requirements.txt
alembic==1.9.4
Bootstrap-Flask==2.2.0
click==8.1.3
dominate==2.7.0
Flask==2.2.3
flask-bootstrap5==0.1.dev1
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
pathlib==1.0.1
# psycopg2-binary==2.9.5
psycopg2
python-dotenv==0.21.1
PyYAML==6.0
SQLAlchemy==2.0.4
supervisor==4.2.5
typing==3.7.4.3
typing_extensions==4.5.0
visitor==0.1.3
Werkzeug==2.2.3
WTForms==3.0.1
